inference_service/requirements.txt
fastapi
uvicorn
torch
rasterio
pyyaml
terratorch
numpy
python-multipart